+++ CHANGELOG.md
@@ -5,9 +5,12 @@ Tous les changements apportés au projet sont suivis dans ce fichier.
 Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0/).
 
 ## [À venir dans la prochaine version]
+### Ajouté
+- Le user puisse voir son graphique de ses temps passés sur les projets personnellement dans le Dashboard
 ### Modifié
 - Style des cartes des actualités dans le tableau de bord
-
+### Réparé
+- Graph de la page utilisateur "2021 Temps passés en heure" /utilisateur/1
 
 
 ## [0.40.12] - 2022-01-18